% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c99.R
\docType{data}
\name{c99}
\alias{c99}
\title{Example of an election of an academic organisation using STV allowing equal preferences frelection om an}
\format{
A list with 10 variables, largest the vote matrix (343 x 10)
\describe{q()
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\usage{
c99
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
