% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc17.R
\docType{data}
\name{cnc17}
\alias{cnc17}
\title{Example of a political STV election - Cumnock and New Cumnock 2017}
\format{
A list with 10 variables, largest the vote matrix (930 x 8)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://www.macs.hw.ac.uk/~denis/stv_elections/SC2017/}
}
\usage{
cnc17
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
