% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc12.R
\docType{data}
\name{hc12}
\alias{hc12}
\title{Unusually simple example of a political STV election - Helensburgh Central 2012}
\format{
A list with 10 variables, largest the vote matrix (629 x 6)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://www.macs.hw.ac.uk/~denis/stv_elections/SC2012/}
}
\usage{
hc12
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
