% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j02.R
\docType{data}
\name{j02}
\alias{j02}
\title{Example of an STV election allowing equal preferences - John Muir Trust 2002}
\format{
A list with 10 variables, largest the vote matrix (1168 x 10)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://www.macs.hw.ac.uk/~denis/stv_elections/jmt2002.blt}
}
\usage{
j02
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
