% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms99.R
\docType{data}
\name{lms99}
\alias{lms99}
\title{Example of an STV election allowing equal preferences - London Mathematical Society 1999}
\format{
A list with 10 variables, largest the vote matrix (343 x 10)
\describe{q()
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\usage{
lms99
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
