% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nws17.R
\docType{data}
\name{nws17}
\alias{nws17}
\title{Example of a political STV election - North West and Central Sutherland 2017
Unusual in that under WIG no candidate reaches the (fixed) quota}
\format{
A list with 10 variables, largest the vote matrix (629 x 6)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://www.macs.hw.ac.uk/~denis/stv_elections/SC2017/}
}
\usage{
nws17
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
