% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlines.R
\docType{data}
\name{outlines}
\alias{outlines}
\title{Lines of html text for use in STV count webpages}
\format{
A character vector of 62 elements, each a line of html text
}
\source{
inst/extdata/outlines.txt
}
\usage{
outlines
}
\description{
Lines of html text for use in STV count webpages
}
\keyword{datasets}
