% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.batch.R
\name{stv.batch}
\alias{stv.batch}
\title{STV election count}
\usage{
stv.batch(
  election,
  indirec,
  outdirec = tempdir(),
  plot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{election}{Name of election}

\item{indirec}{Directory with vote data}

\item{outdirec}{Directory for webpages (if produced)}

\item{plot}{Whether to produce web pages with plots of countcheck()}

\item{verbose}{Whether to report on each stage of the count}
}
\value{
A list containing votes and keep vals at each stage, + optional web pages
}
\description{
STV election count
}
