% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_plots.R
\name{stv.plots}
\alias{stv.plots}
\title{Makes webpage plots of result of an STV election}
\usage{
stv.plots(elecdata, outdirec = tempdir(), webdisplay = TRUE, map = FALSE)
}
\arguments{
\item{elecdata}{An R list of results from stv or stv.wig}

\item{outdirec}{A directory for web page output}

\item{webdisplay}{If TRUE displays the main output web page}

\item{map}{A possible extra showing map of election location}
}
\value{
Webpages with plots of election count and results
}
\description{
Makes webpage plots of result of an STV election
}
\examples{
c99result=stv(c99)
stv.plots(c99result)
nws17wig=stv.wig(nws17)
stv.plots(nws17wig)
}
