% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{STV election count - uses Meek STV, allows equal preferences}
\usage{
stv(
  elecdata,
  outdirec = "out_stv",
  electitle = character(),
  map = F,
  verbose = F,
  plot = T,
  webdisplay = F,
  timing = F
)
}
\arguments{
\item{elecdata}{File with vote data}

\item{outdirec}{Needs to be set for permanent record of results}

\item{electitle}{For web page heading links if appropriate}

\item{map}{Link to a map or other URL associated with election}

\item{verbose}{If =T reports and pauses at each stage of the count
(press return to continue to next stage)}

\item{plot}{If =T (default) produces plots of count and webpages in outdirec}

\item{webdisplay}{If =T displays plots and statistics as web pages}

\item{timing}{Whether to report computing time at each stage}
}
\value{
A list containing votes and keep vals at each stage, + optional web pages; for details see manual pref_pkg_manual.pdf
}
\description{
STV election count - uses Meek STV, allows equal preferences
}
\examples{
hc12m=stv(hc12,plot=FALSE)
j02m=stv(j02,plot=FALSE)
cnc17m=stv(cnc17,plot=FALSE)
y12m=stv(y12,plot=FALSE)

}
