% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_result.R
\name{stv.result}
\alias{stv.result}
\title{Prints main details of result of an STV election}
\usage{
stv.result(elecdata, outfile = "terminal")
}
\arguments{
\item{elecdata}{An R list of results from stv or stv.wig}

\item{outfile}{Where to print results (terminal or file path)}
}
\value{
Prints out results
}
\description{
Prints main details of result of an STV election
}
\examples{
c99result=stv(c99)
stv.result(c99result)
nws17wig=stv.wig(nws17)
stv.result(nws17wig)
}
