% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y12.R
\docType{data}
\name{y12}
\alias{y12}
\title{Example of an STV election - Yale faculty election (last 12)}
\format{
A list with 10 variables, largest the vote matrix (424 x 44)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://cran.r-project.org/package=STV}
}
\usage{
y12
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
