% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yale.R
\docType{data}
\name{yale}
\alias{yale}
\title{Example of an STV election with many candidates - a Yale faculty election}
\format{
A list with 10 variables, largest the vote matrix (479 x 44)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://cran.r-project.org/web/packages/STV/}
}
\usage{
yale
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
