% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{pref.data}
\alias{pref.data}
\title{put election data in an R file (.rda)}
\usage{
pref.data(
  datafile,
  mult = F,
  details = T,
  parties = F,
  ballot = F,
  friendly = F
)
}
\arguments{
\item{datafile}{File with election data}

\item{mult}{Whether includes aggregated votes (default F)}

\item{details}{Whether full election detail (default) or just vote matrix}

\item{parties}{File with party details (default F, i.e. omit)}

\item{ballot}{Default F (meaning pref format)}

\item{friendly}{Default F (meaning most details after votes)}
}
\value{
A standardised list of election info to save in a .rda file
}
\description{
put election data in an R file (.rda)
}
\examples{
datafile=system.file("extdata","Partick2017.blt",package="pref")
parties17=system.file("extdata","parties_SC2017.txt",package="pref")
p17=pref.data(datafile,mult=TRUE,parties=parties17,ballot=TRUE,friendly=TRUE)
}
