% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{stv.data}
\alias{stv.data}
\title{put election data in an R file (.rda)}
\usage{
stv.data(datafile, mult = F, parties = F, format = "ballot", details = T)
}
\arguments{
\item{datafile}{File with election data}

\item{mult}{Whether includes aggregated votes (default F)}

\item{parties}{File with party details (default F, i.e. omit)}

\item{format}{Default "ballot"}

\item{details}{Whether full election detail (default) or just vote matrix}
}
\value{
A standardised list of election info to save in a .rda file
}
\description{
put election data in an R file (.rda)
}
