% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{STV election count}
\usage{
stv(elecdata, outdirec = tempdir(), verbose = T, plot = T)
}
\arguments{
\item{elecdata}{File with vote data}

\item{outdirec}{Directory for webpages (if produced)}

\item{verbose}{Whether to report on each stage of the count}

\item{plot}{Whether to produce web pages with plots of countcheck()}
}
\value{
A list containing votes and keep vals at each stage, + optional web pages
}
\description{
STV election count
}
\examples{
j02c=stv(j02)
p17c=stv(p17)
y12c=stv(y12)

}
