% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{STV election count}
\usage{
stv(
  elecdata,
  outdirec = tempdir(),
  electitle = character(),
  map = F,
  verbose = F,
  plot = T,
  webdisplay = F,
  timing = F
)
}
\arguments{
\item{elecdata}{File with vote data}

\item{outdirec}{Directory for webpages (if produced)}

\item{electitle}{For web page heading links if appropriate}

\item{map}{Link to a map or other URL associated with election}

\item{verbose}{Whether to report on each stage of the count}

\item{plot}{Whether to produceplots of count}

\item{webdisplay}{Whether to display plotsand statistics as web pages}

\item{timing}{Whether to report computing time at each stage}
}
\value{
A list containing votes and keep vals at each stage, + optional web pages
}
\description{
STV election count
}
\examples{
hc12m=stv(hc12,plot=FALSE)
j02m=stv(j02,plot=FALSE)
nws17m=stv(nws17,plot=FALSE)
y12m=stv(y12,plot=FALSE)

}
