% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_result.R
\name{stv.result}
\alias{stv.result}
\title{Prints main details of result of an STV election}
\usage{
stv.result(outlist)
}
\arguments{
\item{outlist}{An R list of results from stv or stv.wig}
}
\value{
Prints out results
}
\description{
Prints main details of result of an STV election
}
\examples{
lms99result=stv(lms99)
stv.result(lms99result)
nws17wig=stv.wig(nws17)
stv.result(nws17wig)
}
