% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.wig.R, R/stv_wig.R
\name{stv.wig}
\alias{stv.wig}
\title{STV election count using WIG as for Scottish Council elections
calculated to 5 places of decimals as used for those elections}
\usage{
stv.wig(
  elecdata,
  outdirec = tempdir(),
  map = FALSE,
  electitle = character(),
  verbose = F,
  plot = F,
  timing = F
)

stv.wig(
  elecdata,
  outdirec = tempdir(),
  map = FALSE,
  electitle = character(),
  verbose = F,
  plot = F,
  timing = F
)
}
\arguments{
\item{elecdata}{File with vote data}

\item{outdirec}{Directory for webpages (if produced)}

\item{map}{Link to a map or other URL associated with election}

\item{electitle}{For web page heading links if appropriate}

\item{verbose}{Whether to report on each stage of the count}

\item{plot}{Whether to produce web pages with plots of countcheck()}

\item{timing}{Whether to report computing time at each stage}
}
\value{
A list containing votes at each stage, + optional web pages

A list containing votes at each stage, + optional web pages
}
\description{
STV election count using WIG as for Scottish Council elections
calculated to 5 places of decimals as used for those elections

STV election count using WIG as for Scottish Council elections
calculated to 5 places of decimals as used for those elections
}
\examples{
p17w=stv.wig(p17)
nws17w=stv.wig(nws17)

nws17w=stv.wig(nws17,plot=FALSE)
p17w=stv.wig(p17,plot=FALSE)

}
