% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_wig.R
\name{stv.wig}
\alias{stv.wig}
\title{STV election count using WIG as for Scottish Council elections
calculated to 5 places of decimals as used for those elections}
\usage{
stv.wig(
  elecdata,
  outdirec = "out_wig",
  electitle = character(),
  map = F,
  verbose = F,
  plot = T,
  webdisplay = F,
  timing = F
)
}
\arguments{
\item{elecdata}{File with vote data}

\item{outdirec}{Needs to be set for permanent record of results}

\item{electitle}{For web page heading links if appropriate}

\item{map}{Link to a map or other URL associated with election}

\item{verbose}{If =T reports and pauses at each stage of the count
(press return to continue to next stage)}

\item{plot}{If =T (default) produces plots of count and webpages in outdirec}

\item{webdisplay}{If =T displays plots and statistics as web pages}

\item{timing}{Whether to report computing time at each stage}
}
\value{
A list containing votes at each stage, + optional web pages; for details see manual pref_pkg_manual.pdf (section 3)
}
\description{
STV election count using WIG as for Scottish Council elections
calculated to 5 places of decimals as used for those elections
}
\examples{
nws17w=stv.wig(nws17)
p17w=stv.wig(p17,plot=FALSE)

}
