% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv_wig.R
\name{stv.wig}
\alias{stv.wig}
\title{STV election count using WIG as for Scottish Council elections}
\usage{
stv.wig(elecdata, verbose = T, plot = T, outdirec = tempdir())
}
\arguments{
\item{elecdata}{File with vote data}

\item{verbose}{Whether to report on each stage of the count}

\item{plot}{Whether to produce web pages with plots of countcheck()}

\item{outdirec}{Directory for webpages (if produced)}
}
\value{
A list containing votes at each stage, + optional web pages
}
\description{
STV election count using WIG as for Scottish Council elections
}
\examples{
p17w=stv.wig(p17)
nws17w=stv.wig(nws17)

}
