% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y12.R
\docType{data}
\name{y12}
\alias{y12}
\title{Example of an STV election allowing equal preferences - John Muir Trust 2002}
\format{
A list with 10 variables, largest the vote matrix (426 x 12)
\describe{
\item{s}{number of seats}
\item{c}{number of candidates}
...
}
}
\source{
\url{https://cran.r-project.org/web/packages/STV/}
}
\usage{
y12
}
\description{
Votes are in a ballot-format matrix
}
\keyword{datasets}
